Move complex Image constructor logic to action
