+++ packages/framework/src/Concerns/FrontMatter/Schemas/BlogPostSchema.php
@@ -2,6 +2,7 @@
 
 namespace Hyde\Framework\Concerns\FrontMatter\Schemas;
 
+use Hyde\Framework\Actions\Constructors\ConfiguresFeaturedImageForPost;
 use Hyde\Framework\Actions\Constructors\FindsAuthorForPost;
 use Hyde\Framework\Models\Author;
 use Hyde\Framework\Models\DateString;
@@ -44,36 +45,6 @@ trait BlogPostSchema
         $this->description = $this->matter('description', substr($this->markdown, 0, 125).'...');
         $this->date = $this->matter('date') !== null ? new DateString($this->matter('date')) : null;
         $this->author = FindsAuthorForPost::run($this);
-        $this->constructImage();
-    }
-
-    private function constructImage(): void
-    {
-        if ($this->matter('image') !== null) {
-            if (is_string($this->matter('image'))) {
-                $this->image = $this->constructBaseImage($this->matter('image'));
-            }
-            if (is_array($this->matter('image'))) {
-                $this->image = $this->constructFullImage($this->matter('image'));
-            }
-        }
-    }
-
-    private function constructBaseImage(string $image): Image
-    {
-        if (str_starts_with($image, 'http')) {
-            return new Image([
-                'uri' => $image,
-            ]);
-        }
-
-        return new Image([
-            'path' => $image,
-        ]);
-    }
-
-    private function constructFullImage(array $image): Image
-    {
-        return new Image($image);
+        $this->image = ConfiguresFeaturedImageForPost::run($this);
     }
 }
